Move text entry test back to its original location after RawKeyEvent migration (Resolves #1700)
